requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
boto3==1.34.0
redis==5.0.1
python-magic==0.4.27
Pillow==10.1.0